+++ packages/react/src/experimental/Information/Avatars/EmojiAvatar/index.tsx
@@ -8,8 +8,8 @@ type Props = {
 
 const sizes = {
   sm: "w-6 h-6 rounded-sm",
-  md: "w-9 h-9 rounded-md",
-  lg: "w-10 h-10 rounded-lg",
+  md: "w-9 h-9 rounded",
+  lg: "w-10 h-10 rounded-md",
 }
 
 const imageSizes = {