+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/index.tsx
@@ -283,7 +283,7 @@ export const TableCollection = <
   const checkColumnWidth = source.selectable ? 52 : 0
 
   return (
-    <div className="flex h-full flex-col gap-4">
+    <div className="flex h-full min-h-0 flex-col gap-4">
       <OneTable loading={isLoading}>
         <TableHeader sticky={true}>
           <TableRow>
@@ -368,7 +368,7 @@ export const TableCollection = <
                       }
                     >
                       <GroupHeader
-                        className="px-4"
+                        className="px-3"
                         selectable={!!source.selectable}
                         select={statusToChecked(
                           groupAllSelectedStatus[group.key]