+++ packages/react/src/lib/providers/i18n/i18n-provider-defaults.ts
@@ -36,6 +36,7 @@ export const defaultTranslations = {
     moveDown: "Move down",
     thumbsUp: "Like",
     thumbsDown: "Dislike",
+    other: "Other actions",
   },
   status: {
     selected: {